Converted collectSubSections to cferdinandi's nextUntil with tests
