+++ src/scrollNav.js
@@ -17,6 +17,13 @@ import teardownScrollHandler from './teardownScrollHandler';
 let clickHandler;
 let scrollHandler;
 
+// matches() polyfill
+if (!Element.prototype.matches) {
+  Element.prototype.matches =
+    Element.prototype.msMatchesSelector ||
+    Element.prototype.webkitMatchesSelector;
+}
+
 function isElement(element) {
   return element instanceof Element;
 }
